add JWT RBAC to secure internal rest calls

+++ javaee-cafe/src/main/java/cafe/web/rest/CafeResource.java
@@ -1,12 +1,7 @@
 package cafe.web.rest;
 
-import java.lang.invoke.MethodHandles;
 import java.util.List;
-import java.util.logging.Level;
-import java.util.logging.Logger;
-
 import javax.inject.Inject;
-import javax.persistence.PersistenceException;
 import javax.ws.rs.Consumes;
 import javax.ws.rs.DELETE;
 import javax.ws.rs.GET;
@@ -18,16 +13,24 @@ import javax.ws.rs.WebApplicationException;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 
+import org.eclipse.microprofile.config.inject.ConfigProperty;
+import org.eclipse.microprofile.jwt.JsonWebToken;
+
 import cafe.model.CafeRepository;
 import cafe.model.entity.Coffee;
 
 @Path("coffees")
 public class CafeResource {
 
-	private static final Logger logger = Logger.getLogger(MethodHandles.lookup().lookupClass().getName());
-
 	@Inject
 	private CafeRepository cafeRepository;
+
+	@Inject
+    @ConfigProperty(name = "admin.group.id")
+    private String ADMIN_GROUP_ID;
+
+    @Inject
+    private JsonWebToken jwtPrincipal;
 
 	@GET
 	@Produces({ MediaType.APPLICATION_XML, MediaType.APPLICATION_JSON })
@@ -39,12 +42,7 @@ public class CafeResource {
 	@Consumes({ MediaType.APPLICATION_XML, MediaType.APPLICATION_JSON })
 	@Produces({ MediaType.APPLICATION_XML, MediaType.APPLICATION_JSON })	
 	public Coffee createCoffee(Coffee coffee) {
-		try {
-			return this.cafeRepository.persistCoffee(coffee);
-		} catch (PersistenceException e) {
-			logger.log(Level.SEVERE, "Error creating coffee {0}: {1}.", new Object[] { coffee, e });
-			throw new WebApplicationException(e, Response.Status.INTERNAL_SERVER_ERROR);
-		}
+		return this.cafeRepository.persistCoffee(coffee);
 	}
 
 	@GET
@@ -57,12 +55,10 @@ public class CafeResource {
 	@DELETE
 	@Path("{id}")
 	public void deleteCoffee(@PathParam("id") Long coffeeId) {
-		try {
-			this.cafeRepository.removeCoffeeById(coffeeId);
-		} catch (IllegalArgumentException ex) {
-			logger.log(Level.SEVERE, "Error calling deleteCoffee() for coffeeId {0}: {1}.",
-					new Object[] { coffeeId, ex });
-			throw new WebApplicationException(Response.Status.NOT_FOUND);
-		}
+        if (!this.jwtPrincipal.getGroups().contains(ADMIN_GROUP_ID)) {
+            throw new WebApplicationException(Response.Status.FORBIDDEN);
+        }
+        
+		this.cafeRepository.removeCoffeeById(coffeeId);
 	}
 }
